# docs/api/psyche-loom-openapi.yaml
openapi: 3.1.0
info:
  title: Psyche Loom API
  description: |
    RESTful API for memory management with text-based memory fragments, metadata support, and view-based access patterns.

    ## Core Design Principles

    Psyche Loom focuses on efficient memory management and retrieval:
    - **View-Based Design**: Structured access via `/views/*` endpoints
    - **Flexible Metadata**: JSON metadata for rich context storage
    - **Time-based Queries**: Timeline view for time range queries
    - **Clean Architecture**: Simple RESTful design with consistent responses
  version: 1.0.0
  contact:
    name: Psyche Loom Team
    email: team@psyche-loom.example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT

servers:
  - url: http://127.0.0.1:8080
    description: Development server

paths:
  /health:
    get:
      summary: Health Check
      description: Check if the service is running and healthy
      operationId: healthCheck
      tags:
        - Health
      responses:
        '200':
          description: Service is healthy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
        '500':
          $ref: '#/components/responses/InternalError'

  /api/v1/memories:
    post:
      summary: Create Memories
      description: Store new memory fragments with content and optional metadata (supports batch operations)
      operationId: createMemories
      tags:
        - Memories
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMemoryRequest'
      responses:
        '200':
          description: Memories created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'

  /api/v1/memories/{id}:
    get:
      summary: Get Memory
      description: Retrieve a specific memory fragment by ID
      operationId: getMemory
      tags:
        - Memories
      parameters:
        - name: id
          in: path
          required: true
          description: Memory ID
          schema:
            type: integer
            format: int64
          example: 123
      responses:
        '200':
          description: Memory retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

    delete:
      summary: Delete Memory
      description: Remove a memory fragment by ID
      operationId: deleteMemory
      tags:
        - Memories
      parameters:
        - name: id
          in: path
          required: true
          description: Memory ID
          schema:
            type: integer
            format: int64
          example: 123
      responses:
        '200':
          description: Memory deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

  /api/v1/memories/views/recent:
    get:
      summary: Recent Memories View
      description: Get the most recent memory fragments
      operationId: getRecentMemories
      tags:
        - Views
      parameters:
        - name: limit
          in: query
          required: true
          description: Maximum number of memories to return
          schema:
            type: integer
            minimum: 1
            maximum: 100
          example: 10
      responses:
        '200':
          description: Recent memories retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'

  /api/v1/memories/views/timeline:
    get:
      summary: Timeline View
      description: Get memory fragments within a specific time range
      operationId: getTimelineMemories
      tags:
        - Views
      parameters:
        - name: from
          in: query
          required: true
          description: Start time in ISO 8601 format
          schema:
            type: string
            format: date-time
          example: "2024-01-01T00:00:00Z"
        - name: to
          in: query
          required: true
          description: End time in ISO 8601 format
          schema:
            type: string
            format: date-time
          example: "2024-12-31T23:59:59Z"
        - name: limit
          in: query
          required: false
          description: Maximum number of memories to return
          schema:
            type: integer
            minimum: 1
          example: 50
        - name: offset
          in: query
          required: false
          description: Number of memories to skip for pagination
          schema:
            type: integer
            minimum: 0
          example: 0
      responses:
        '200':
          description: Timeline memories retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'

components:
  schemas:
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
          description: Whether the operation was successful
        data:
          description: Response data (present on success)
          $ref: '#/components/schemas/MemoryResponse'
        error:
          type: string
          description: Error message (present on failure)
          nullable: true
      required:
        - success
      example:
        success: true
        data:
          fragments:
            - id: 123
              content: "Example memory"
              metadata: {}
              source:
                channel: "user"
                identifier: "input"
              timestamp: "2024-01-15T10:30:00Z"
          total: 1
        error: null

    CreateMemoryRequest:
      type: object
      properties:
        fragments:
          type: array
          items:
            $ref: '#/components/schemas/MemoryFragmentInput'
          description: Array of memory fragments to create
      required:
        - fragments
      example:
        fragments:
          - content: "Meeting notes about project architecture"
            source:
              channel: "meeting"
              identifier: "user_input"

    MemoryFragmentInput:
      type: object
      properties:
        content:
          type: string
          description: Main content of the memory fragment
          example: "Meeting notes about project architecture"
        metadata:
          type: object
          description: Optional JSON metadata for additional context
          nullable: true
          example:
            type: "meeting"
            project: "loom"
            participants: ["alice", "bob"]
        source:
          $ref: '#/components/schemas/Source'
      required:
        - content
        - source

    MemoryResponse:
      type: object
      properties:
        fragments:
          type: array
          items:
            $ref: '#/components/schemas/MemoryFragment'
          description: List of memory fragments
        total:
          type: integer
          description: Total number of fragments
          example: 1
      required:
        - fragments
        - total

    MemoryFragment:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Unique memory identifier
          example: 123
        content:
          type: string
          description: Main content of the memory
          example: "Meeting notes about project architecture"
        metadata:
          type: object
          description: JSON metadata for additional context
          nullable: true
          example:
            type: "meeting"
            project: "loom"
        source:
          $ref: '#/components/schemas/Source'
        timestamp:
          type: string
          format: date-time
          description: Creation timestamp (ISO 8601)
          example: "2024-01-15T10:30:00Z"
      required:
        - id
        - content
        - source
        - timestamp

    Source:
      type: object
      properties:
        channel:
          type: string
          description: Source channel (e.g., "meeting", "chat", "note")
          example: "meeting"
        identifier:
          type: string
          description: Source identifier
          example: "user_input"
      required:
        - channel
        - identifier

    HealthResponse:
      type: object
      properties:
        status:
          type: string
          description: Service health status
          example: "ok"
        service:
          type: string
          description: Service name
          example: "loom-memory-service"
        timestamp:
          type: integer
          format: int64
          description: Response timestamp (Unix timestamp)
          example: 1640995200
      required:
        - status
        - service
        - timestamp

    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        data:
          nullable: true
          example: null
        error:
          type: string
          description: Error description
          example: "Invalid request parameters"
      required:
        - success
        - error

  responses:
    BadRequest:
      description: Bad request - invalid parameters or malformed request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            data: null
            error: "Missing required field: content"

    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            data: null
            error: "Memory with ID 123 not found"

    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            data: null
            error: "Database connection failed"

tags:
  - name: Memories
    description: Memory fragment operations
  - name: Views
    description: View-based access patterns (recent, timeline)
  - name: Health
    description: Service health check